requirements.txt
tqdm==4.64.0
requests==2.28.1
beautifulsoup4==4.11.1
transformers==4.21.0
torch==1.12.0
Levenshtein==0.20.1
sentence-splitter==1.4
spacy==3.4.1
nltk==3.7
scikit-learn==1.1.1
evaluate==0.2.2
scipy==1.9.0
laserembeddings==1.1.2
torchmetrics==0.9.3
tensorflow-gpu==2.9.1
numpy~=1.23.1
xlrd~=2.0.1
pandas~=1.5.0
openpyxl~=3.0.10
rouge-score~=0.1.2
rouge==1.0.0
pingouin==0.5.2